move media types and db under same folder
