+++ backend/src/database/media/db.media_crop.service.ts
@@ -1,13 +1,13 @@
 import { BadRequestException, Injectable } from "@nestjs/common";
-import { DbService } from "./db.service";
+import { DbService } from "../db.service";
 import {
   CreateMediaCropDto,
   MediaCropDto,
   ModifyMediaCropDto,
   PaginationFilterDto,
   ReplaceMediaCropDto,
-} from "../dto/dto";
-import { ResourceGoneException } from "../common/exceptions";
+} from "../../dto/dto";
+import { ResourceGoneException } from "../../common/exceptions";
 import { PaginatedResponse } from "@repo/common";
 
 @Injectable()
