+++ backend/src/database/media/db.media_gallery.service.ts
@@ -1,13 +1,13 @@
 import { BadRequestException, Injectable } from "@nestjs/common";
-import { DbService } from "./db.service";
+import { DbService } from "../db.service";
 import {
   CreateMediaGalleryDto,
   MediaGalleryDto,
   MediaItemDto,
   ModifyMediaGalleryDto,
   ReplaceMediaGalleryDto,
-} from "../dto/dto";
-import { ResourceGoneException } from "../common/exceptions";
+} from "../../dto/dto";
+import { ResourceGoneException } from "../../common/exceptions";
 import { PaginatedResponse } from "@repo/common";
 
 @Injectable()
