+++ backend/src/database/media/db.media_item.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Injectable } from "@nestjs/common";
-import { DbService } from "./db.service";
+import { DbService } from "../db.service";
 import {
   CreateMediaItemDto,
   MediaCropDto,
@@ -7,9 +7,9 @@ import {
   ModifyMediaItemDto,
   PaginationFilterDto,
   ReplaceMediaItemDto,
-} from "../dto/dto";
-import { ResourceGoneException } from "../common/exceptions";
-import { PaginatedResponse } from "@repo/common/src/objects/pagination";
+} from "../../dto/dto";
+import { ResourceGoneException } from "../../common/exceptions";
+import { PaginatedResponse } from "common/src/objects/pagination";
 
 @Injectable()
 export class MediaItemDatabaseService {
